# Optimize StreamingResponseConverter