… tests/Tinkoff.Trading.OpenApi.Tests/Json/streaming-candle-response-with-new-field.json