Оптимизация сериализации streaming объектов.
Аллокация памяти в ~4 раза меньше. Производительность возросла на ~30%.

|   Method |     Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
|      Old | 7.994 us | 0.0849 us | 0.0794 us |  1.00 | 0.6104 |     - |     - |    1920 B |
| Improved | 5.633 us | 0.0381 us | 0.0357 us |  0.70 | 0.1602 |     - |     - |     512 B |